pyproject.toml
[tool.poetry]
name = "async-db-polars"
version = "0.1.6"
description = "High level DB abstraction that uses polars to represent tables in memory."
authors = ["WolfEYc <wolfeycode@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
polars = "^0.20.3"
asyncpg = "^0.29.0"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.28.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
